Page 2.html
<html>
    <title>Story Time</title>
    <body>
        <h1>True Conservation</h1>
      
<p>To want to be part of something more than ourselves is always on our mind as we walk through our 
day-to-day lives.  We join teams because we want to live for our brothers and sisters beside us. Some 
of us think we can enhance and elevate what we are working for as the collective. Others want to live 
for another.  There are countless examples of brave men and women that have done, and will continue to 
do, acts to protect others. Some they have met and are very fond of, others they may never know and 
will never have the chance of knowing.But, they know one thing for certain the collective is stronger, 
better, and more appealing than what we can ever have ourselveson a singular level.  </p>

<p>It is this natural born sense of wanting to be part of something better, and create something better, 
that makes me want to be a hunter.  The hunter has lived through the ages as the hero. Hunting parties 
would come home and the villages would gather around to greet them, thankful for their bountiful harvests. 
When I think about the hunter, I think of an ancient or even prehistoric being that provided. They saw 
past their current position at the time.  They could never have envisioned they would be the people that 
would set me on the trail towards conservation of sustainable resources.  All they were doing was trying 
to survive.  Did they know that if they over-harvested, or over-utilized a resource that they would not 
have it anymore for future generations? Did the thought even cross their mind? Or were they living in 
the moment? I believe that they learnedthrough trial and error that if they over-utilized a species of 
animal, or over-harvested a type of plant, that they would no longer have it. </p>

<p>There is the school of thought that ancient humans would have never thought about conservation.  If you 
look at the historical evidence, it is plausible that there was no thought about conservation of utilizable 
resources.  Native Americans ran large amounts of buffalo off of cliffs and hundreds would die at a time.  
There are historical accounts of colonial explorers crossing the lands encountering the wretched stench 
of stained air for miles from the carcasses of unused buffalo meat.  We have this iconic idea that everything 
was utilized, unfortunately, this iconic idea doesn’t make any logical sense.  They did what they had to 
do to survive.  They did what they could to cultivate and harvest the most calorie-efficient diet.  Is it 
their fault? How could they possibly have known that they would have driven one of their main resources, 
the buffalo, to a point that they were on a slide to what would eventually be near extinction? Of course, 
they were not the only cause for the drive of the buffalo to their near extinction.  The market hunters 
had no regard or thought that they would not rebound and have a devastating part in the buffalo’s near extinction. 
​We as a people have learned since that time.  We still have a lot to learn, but we as hunters are on the 
right path towards conserving our natural resources. </p>  

<p> Conservationists are not just the ones that sustainably harvest.  Conservationists are all of us in the 
collective that believe that we need our animal, plant, and other national resources.  You do not have to be a 
hunter to want to conserve the beautiful things that this world has to offer, but you do have to be willing to 
acknowledge that we are the only thing that can wreck wildlife’s existence if we do not do something. We need to 
stop arguing over whether the sustainable harvest of a wild animal based on biological and empirical evidence is 
going to send us crashing into a life of suffering with nothing on our planet.  The true conservation that we need 
to take heed to is the conservation of our lands, waterways, and atmosphere. </p>

<p> On an equal but separate note, there is a lot to be said about the clean air and water that we utilize.  When our 
oceans are filled with garbage, our lands are covered in asphalt, and our air is filled with smoke to the extent 
that cancer spreads to a point of seemingly no return, then it is time that we rethink how we are living our lives.  
We as a collective live no better than those Native Americans or market hunters that decimated the buffalo.  In fact, 
it could be said that we are actually worse because of our separated ideals we believe simply because of an idiotic 
longing to be right.  True conservation concentrates on what we are doing to poison the habitat of the animals we 
love and not to mention ourselves.  </p>

<p> As a hunter, I have learned this.  I have learned thatconservation is not as simple as just sustainable utilization 
of an animal.  It is sustainable utilization of all of our resources that we as a collective take in part.  We must 
stop with the dividing lines.  If we are to continue life for future generations then, we,the non-hunters and hunters, 
have to come together in this true metric that is undoubtedly the true meaning of conservation.</p>
    </body>
   
    <img src="Images/IMG_0105.jpeg"/>

    <p><a href="webpage.html" target="_new">Page 1</a></p>
    <p><a href="Page 3.html" target="_new">Page 3</a></p>
    </html>